+++ src/main.c
@@ -11,7 +11,7 @@
 
 DMM_MallocHeader *dmm_global_instance = DMM_UNASSIGNED_REGION;
 
-void dmm_add_memory_region(void *start, uint64_t length)
+void dmm_add_memory_region(void *start, size_t length)
 {
     void *result = dmm_instance_add_memory_region(dmm_global_instance, start, length);
